feat(DnD): #MAG-484 dnd on free layout
