+++ frontend/src/components/cards-free-layout/CardsFreeLayout.tsx
@@ -5,8 +5,8 @@ import { rectSortingStrategy, SortableContext } from "@dnd-kit/sortable";
 import { Box } from "@mui/material";
 
 import { LiWrapper, UlWrapper, mainWrapperProps } from "./style";
-import { useFreeLayoutCardDnD } from "../../hooks/dnd-hooks/useFreeLayoutCardDnD";
 import { BoardCard } from "../board-card/BoardCard";
+import { useFreeLayoutCardDnD } from "~/hooks/dnd-hooks/useFreeLayoutCardDnD";
 import { useBoard } from "~/providers/BoardProvider";
 
 export const CardsFreeLayout: FC = () => {
